+++ src/main/java/io/contek/invoker/binancelinear/api/common/_FutureDetails.java
@@ -1,17 +1,16 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _FutureDetails {
 
     public String symbol;
-    public BigDecimal markPrice;
-    public BigDecimal indexPrice;
-    public BigDecimal estimatedSettlePrice;
-    public BigDecimal lastFundingRate;
+    public Double markPrice;
+    public Double indexPrice;
+    public Double estimatedSettlePrice;
+    public Double lastFundingRate;
     public long nextFundingTime;
-    public BigDecimal interestRate;
+    public Double interestRate;
     public long time;
 }
