+++ src/main/java/io/contek/invoker/binancelinear/api/common/_LeverageBracket.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _LeverageBracket {
@@ -10,6 +9,6 @@ public class _LeverageBracket {
   public int initialLeverage;
   public long notionalCap;
   public long notionalFloor;
-  public BigDecimal maintMarginRatio;
-  public BigDecimal cum;
+  public Double maintMarginRatio;
+  public Double cum;
 }
