+++ src/main/java/io/contek/invoker/binancelinear/api/common/_UserTrade.java
@@ -1,21 +1,20 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _UserTrade {
 
   public boolean buyer;
-  public BigDecimal commission;
+  public Double commission;
   public String commissionAsset;
   public String id;
   public boolean maker;
   public String orderId;
-  public BigDecimal price;
-  public BigDecimal qty;
-  public BigDecimal quoteQty;
-  public BigDecimal realizedPnl;
+  public Double price;
+  public Double qty;
+  public Double quoteQty;
+  public Double realizedPnl;
   public String side;
   public String positionSide;
   public String symbol;
