+++ src/main/java/io/contek/invoker/binancelinear/api/websocket/user/OrderUpdateChannel.java
@@ -6,7 +6,6 @@ import io.contek.invoker.binancelinear.api.websocket.user.constants.UserEventTyp
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
-import java.math.BigDecimal;
 
 @ThreadSafe
 public final class OrderUpdateChannel extends UserWebSocketChannel<OrderUpdateChannel.Data> {
@@ -45,31 +44,31 @@ public final class OrderUpdateChannel extends UserWebSocketChannel<OrderUpdateCh
     public String S; // side
     public String o; // order type
     public String f; // time in force
-    public BigDecimal q; // original quantity;
-    public BigDecimal p; // original price
-    public BigDecimal ap; // average price
-    public BigDecimal sp; // stop price
+    public Double q; // original quantity;
+    public Double p; // original price
+    public Double ap; // average price
+    public Double sp; // stop price
     public String x; // execution type
     public String X; // order status
     public String i; // order id
-    public BigDecimal l; // order last filled quantity
-    public BigDecimal z; // order filled accumulated quantity
-    public BigDecimal L; // Last Filled Price
+    public Double l; // order last filled quantity
+    public Double z; // order filled accumulated quantity
+    public Double L; // Last Filled Price
     public String N; // commission asset
-    public BigDecimal n; // commission
+    public Double n; // commission
     public Long T; // order trade time
     public String t; // trade id
-    public BigDecimal b; // bids notional
-    public BigDecimal a; // ask notional
+    public Double b; // bids notional
+    public Double a; // ask notional
     public Boolean m; // is the trade the maker side
     public Boolean R; // is this trade reduce only
     public String wt; // stop work price type
     public String ot; // original order type
     public String ps; // position side
     public Boolean cp; // if close all push with conditional order
-    public BigDecimal AP; // activation price
-    public BigDecimal cr; // callback rate
-    public BigDecimal rp; // realized profit
+    public Double AP; // activation price
+    public Double cr; // callback rate
+    public Double rp; // realized profit
     public long traceNano = System.nanoTime();
   }
 }
